ACCELEROMETER AND GEOFENCING POSSIBLY WORKING OMG

+++ Prototype/app/src/main/Java/Models/GeoReport.java
@@ -6,8 +6,8 @@ package Models;
 
 public class GeoReport {
     private String ID;
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
 
     public GeoReport() {
     }
@@ -22,19 +22,19 @@ public class GeoReport {
         return ID;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
